refactor(framework): one content-type table for both static servers
